fix: move SSR component to subpath export
