--- src/client/ssr/index.ts
@@ -1,2 +1,0 @@
-export * from "./SSR";
-export { default } from "./SSR";
