+++ src/ssr/index.ts
@@ -1,0 +1,2 @@
+export type { SSRProps as UnstableSSRProps } from "./SSR";
+export { default as UnstableSSR } from "./SSR";
