+++ test/SSR.spec.tsx
@@ -1,4 +1,5 @@
-import { RowsPhotoAlbum, UnstableSSR as SSR } from "../src";
+import { RowsPhotoAlbum } from "../src";
+import { UnstableSSR as SSR } from "../src/ssr";
 import { render } from "./test-utils";
 import photos from "./photos";
